25/08/amap_async_mongo/pyproject.toml
[project]
name = "amap-async-mongo"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "aiolimiter>=1.2.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pymongo>=4.14.0",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
    "transformers>=4.55.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]

[project.scripts]
amap_insert = "amap_mongo.ali:insert2mongo"
amap_export = "amap_mongo.ali:export2excel"